refactor: Move `get_enabled_mods` to `main.rs`
Part of #329
